refactor: ♻️  拆分 config-provider 主题变量类型生成文件
