+++ scripts/buildThemeVars.ts
@@ -12,7 +12,7 @@ type ThemeVarGroup = {
   fields: ThemeVarField[]
 }
 
-const tsFilePath = path.resolve(__dirname, '../src/uni_modules/wot-ui/components/wd-config-provider/types.ts')
+const themeVarsFilePath = path.resolve(__dirname, '../src/uni_modules/wot-ui/components/wd-config-provider/theme-vars.ts')
 const globalScssPath = path.resolve(__dirname, '../src/uni_modules/wot-ui/styles/variable.scss')
 const componentsRootPath = path.resolve(__dirname, '../src/uni_modules/wot-ui/components')
 
@@ -182,37 +182,12 @@ const renderTypeBlock = (typeName: string, fields: ThemeVarField[]) => {
   return `export type ${typeName} = {\n${lines.join('\n')}\n}\n`
 }
 
-const generateTSFileContent = (baseFields: ThemeVarField[], componentGroups: ThemeVarGroup[]) => {
-  let tsContent = `import type { ExtractPropTypes, PropType, InjectionKey, ComputedRef } from 'vue'
-import { makeStringProp, baseProps } from '../../common/props'
+const generateThemeVarsFileContent = (baseFields: ThemeVarField[], componentGroups: ThemeVarGroup[]) => {
+  let tsContent = `/* eslint-disable quotes */
+// This file is generated by scripts/buildThemeVars.ts.
 
-export const configProviderProps = {
-  ...baseProps,
-  /**
-   * 主题风格，设置为 dark 来开启深色模式，全局生效
-   */
-  theme: makeStringProp('light'),
-  /**
-   * 自定义主题变量
-   */
-  themeVars: {
-    type: Object as PropType<ConfigProviderThemeVars>,
-    default: () => ({})
-  }
-}
-
-export type ConfigProviderProps = ExtractPropTypes<typeof configProviderProps>
-
-export type ConfigProviderProvide = {
-  theme: ComputedRef<string>
-  themeStyle: ComputedRef<string>
-}
-
-export const CONFIG_PROVIDER_KEY: InjectionKey<ConfigProviderProvide> = Symbol('wd-config-provider')
-
+${renderTypeBlock('baseThemeVars', baseFields)}
 `
-
-  tsContent += `${renderTypeBlock('baseThemeVars', baseFields)}\n`
 
   componentGroups.forEach((group) => {
     tsContent += `${renderTypeBlock(group.typeName, group.fields)}\n`
@@ -230,8 +205,13 @@ export const CONFIG_PROVIDER_KEY: InjectionKey<ConfigProviderProvide> = Symbol('
 
 const baseFields = buildBaseThemeVars()
 const componentGroups = buildComponentThemeVars()
-const tsContent = generateTSFileContent(baseFields, componentGroups)
 
-fs.writeFileSync(tsFilePath, tsContent)
+if (!baseFields.length || !componentGroups.length) {
+  throw new Error(`Theme vars generation failed: base=${baseFields.length}, components=${componentGroups.length}`)
+}
+
+const tsContent = generateThemeVarsFileContent(baseFields, componentGroups)
+
+fs.writeFileSync(themeVarsFilePath, tsContent)
 
 console.log(`TS file generated successfully! base: ${baseFields.length}, components: ${componentGroups.length}`)
