+++ src/uni_modules/wot-ui/index.ts
@@ -3,4 +3,4 @@ export { zodAdapter } from './components/wd-form'
 export type { FormSchema, FormSchemaIssue, FormValidateEvent, FormValidateTrigger } from './components/wd-form'
 export * as CommonUtil from './common/util'
 export * from './locale'
-export type { ConfigProviderThemeVars } from './components/wd-config-provider/types'
+export type { ConfigProviderThemeVars } from './components/wd-config-provider/theme-vars'
